Creating Factories for Database Seeding. Fixed models and migration for seeding

+++ database/migrations/2020_04_26_124924_create_products_table.php
@@ -21,7 +21,7 @@ class CreateProductsTable extends Migration
             $table->integer('quantity')->unsigned();
             $table->string('status')->default(ProductStatus::UNAVAILABLE);
             $table->string('image');
-            $table->integer('seller_id')->unsigned();
+            $table->bigInteger('seller_id')->unsigned();
             $table->timestamps();
         });
     }
